+++ cmd/gazelle/BUILD.bazel
@@ -21,10 +21,10 @@ go_library(
         "//internal/language/go:go_default_library",
         "//internal/language/proto:go_default_library",
         "//internal/merger:go_default_library",
-        "//internal/repos:go_default_library",
         "//internal/version:go_default_library",
         "//internal/walk:go_default_library",
         "//label:go_default_library",
+        "//repos:go_default_library",
         "//resolve:go_default_library",
         "//rule:go_default_library",
         "//vendor/github.com/pmezard/go-difflib/difflib:go_default_library",
