+++ internal/language/go/BUILD.bazel
@@ -50,9 +50,9 @@ go_library(
         "//internal/flag:go_default_library",
         "//internal/language:go_default_library",
         "//internal/language/proto:go_default_library",
-        "//internal/repos:go_default_library",
         "//label:go_default_library",
         "//pathtools:go_default_library",
+        "//repos:go_default_library",
         "//resolve:go_default_library",
         "//rule:go_default_library",
         "//vendor/github.com/bazelbuild/buildtools/build:go_default_library",
@@ -76,10 +76,10 @@ go_test(
         "//internal/language:go_default_library",
         "//internal/language/proto:go_default_library",
         "//internal/merger:go_default_library",
-        "//internal/repos:go_default_library",
         "//internal/testtools:go_default_library",
         "//internal/walk:go_default_library",
         "//label:go_default_library",
+        "//repos:go_default_library",
         "//resolve:go_default_library",
         "//rule:go_default_library",
         "//vendor/github.com/bazelbuild/buildtools/build:go_default_library",
